+++ chatterplot/app/src/main/res/layout/dataset_card.xml
@@ -31,14 +31,6 @@
             android:textSize="24sp"
             android:textStyle="bold" />
 
-        <ImageButton
-            android:id="@+id/deleteDatasetButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentStart="true"
-            android:layout_alignParentTop="true"
-            app:srcCompat="@android:drawable/ic_menu_delete" />
-
     </RelativeLayout>
